attempts to better separate form function
