+++ src/helperFunctions/formHelper.ts
@@ -1,6 +1,5 @@
 /* eslint-disable no-case-declarations */
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { licences } from "../mockData/licences";
 import {
   DataTypeStep,
   BenefitsStep,
@@ -21,7 +20,233 @@ import {
   FormData,
   GenericStringArray,
 } from "../types/express";
+import { licences } from "../mockData/licences";
 import { replace } from "./checkhelper";
+
+export const updateStepsStatus = (
+  currentStep: string,
+  stepValue: StepValue,
+  formdata: FormData,
+  returnToStart: boolean,
+) => {
+  const completedSections = new Set();
+
+  // Group up the steps so we can work out which sections have been completed later
+  const purposeSteps = [
+    "data-type",
+    "data-subjects",
+    "project-aims",
+    "data-required",
+    "benefits",
+    "data-access",
+    "other-orgs",
+    "impact",
+    "date",
+  ];
+
+  const legalSteps = [
+    "legal-power",
+    "legal-power-advice",
+    "legal-gateway",
+    "legal-gateway-advice",
+  ];
+
+  const dataProtectionSteps = [
+    "lawful-basis-personal",
+    "lawful-basis-special",
+    "lawful-basis-special-public-interest",
+    "data-travel",
+    "data-travel-location",
+    "role",
+  ];
+
+  const securitySteps = ["delivery", "format", "disposal"];
+
+  // If "Save and return" was clicked, set to "IN PROGRESS" if needed and return
+  if (returnToStart) {
+    if (formdata.steps[currentStep].status !== "COMPLETED") {
+      formdata.steps[currentStep].status = "IN PROGRESS";
+    }
+    return;
+  } else {
+    // If "Save and Continue" was clicked, set this step to "COMPLETED"
+    formdata.steps[currentStep].status = "COMPLETED";
+  }
+
+  if (currentStep === "data-type") {
+    const notRequiredSteps = new Set<string>();
+    const notStartedSteps = new Set<string>();
+    const val = stepValue as DataTypeStep;
+    // If personal is not checked then lawful-basis-personal is not required
+    if (!val.personal.checked) {
+      notRequiredSteps.add("lawful-basis-personal");
+    } else {
+      notStartedSteps.add("data-subjects");
+      notStartedSteps.add("lawful-basis-personal");
+      notStartedSteps.add("role");
+    }
+    // If special is not checked then lawful-basis-special is not required
+    if (!val.special.checked) {
+      notRequiredSteps.add("lawful-basis-special");
+      notRequiredSteps.add("lawful-basis-special-public-interest");
+    } else {
+      notStartedSteps.add("data-subjects");
+      notStartedSteps.add("lawful-basis-special");
+      notStartedSteps.add("role");
+    }
+    // If none is checked, a few other steps are not required
+    if (val.none.checked) {
+      [
+        "data-subjects",
+        "lawful-basis-personal",
+        "lawful-basis-special",
+        "lawful-basis-special-public-interest",
+        "role",
+      ].forEach((s) => notRequiredSteps.add(s));
+    }
+
+    // Set everything that's not required to NOT REQUIRED
+    for (const s of notRequiredSteps) {
+      formdata.steps[s].status = "NOT REQUIRED";
+    }
+
+    // Set everything that needs to be completed to NOT STARTED
+    for (const s of notStartedSteps) {
+      const stepStatus = formdata.steps[s].status;
+      if (!["COMPLETED", "IN PROGRESS"].includes(stepStatus)) {
+        formdata.steps[s].status = "NOT STARTED";
+      }
+    }
+  }
+
+  if (currentStep === "data-access") {
+    if (!stepValue || stepValue === "no") {
+      formdata.steps["other-orgs"].status = "NOT REQUIRED";
+    } else {
+      formdata.steps["other-orgs"].status = "NOT STARTED";
+    }
+  }
+
+  if (currentStep === "legal-power") {
+    if ((stepValue as LegalPowerStep).yes.checked) {
+      formdata.steps["legal-power-advice"].status = "NOT REQUIRED";
+    } else {
+      formdata.steps["legal-power-advice"].status = "NOT STARTED";
+      formdata.steps[currentStep].status = "IN PROGRESS";
+    }
+  }
+
+  if (currentStep === "legal-gateway") {
+    const legalGatewayStep = stepValue as LegalGatewayStep;
+    if (legalGatewayStep.yes.checked || legalGatewayStep.other.checked) {
+      formdata.steps["legal-gateway-advice"].status = "NOT REQUIRED";
+    } else {
+      formdata.steps["legal-gateway-advice"].status = "NOT STARTED";
+      formdata.steps[currentStep].status = "IN PROGRESS";
+    }
+  }
+
+  if (currentStep === "check") {
+    formdata.steps["declaration"].status = "NOT STARTED";
+  }
+
+  if (currentStep === "declaration") {
+    formdata.steps["confirmation"].status = "NOT STARTED";
+  }
+
+  if (currentStep === "lawful-basis-special") {
+    if (
+      (stepValue as LawfulBasisSpecialStep)["reasons-of-public-interest"]
+        ?.checked
+    ) {
+      formdata.steps["lawful-basis-special-public-interest"].status =
+        "NOT STARTED";
+    } else {
+      formdata.steps["lawful-basis-special-public-interest"].status =
+        "NOT REQUIRED";
+    }
+  }
+
+  if (currentStep === "data-travel-location") {
+    formdata.steps["data-travel"].status = "COMPLETED";
+  }
+
+  if (currentStep === "data-travel") {
+    if (!stepValue || stepValue === "no") {
+      formdata.steps["data-travel-location"].status = "NOT REQUIRED";
+    } else {
+      formdata.steps["data-travel-location"].status = "NOT STARTED";
+      formdata.steps[currentStep].status = "IN PROGRESS";
+    }
+  }
+
+  // Loop over all the steps in each section to check whether the
+  //  section is complete and/or the 'check' step can be enabled
+  if (everyStepCompleted(purposeSteps, formdata)) {
+    completedSections.add("purpose");
+  } else {
+    completedSections.delete("purpose");
+  }
+
+  if (everyStepCompleted(legalSteps, formdata)) {
+    // If all the legal steps AND the legal review is completed, legal is done.
+    if (formdata.steps["legal-review"].status === "COMPLETED") {
+      completedSections.add("legal");
+    } else {
+      formdata.steps["legal-review"].status = "NOT STARTED";
+    }
+  } else {
+    // If not all of the legal steps are Completed, legal review cannot be started
+    formdata.steps["legal-review"].status = "CANNOT START YET";
+    completedSections.delete("legal");
+  }
+
+  if (everyStepCompleted(dataProtectionSteps, formdata)) {
+    if (formdata.steps["protection-review"].status === "COMPLETED") {
+      completedSections.add("data-protection");
+    } else {
+      formdata.steps["protection-review"].status = "NOT STARTED";
+    }
+  } else {
+    formdata.steps["protection-review"].status = "CANNOT START YET";
+    completedSections.delete("data-protection");
+  }
+
+  if (everyStepCompleted(securitySteps, formdata)) {
+    if (formdata.steps["security-review"].status === "COMPLETED") {
+      completedSections.add("security");
+    } else {
+      formdata.steps["security-review"].status = "NOT STARTED";
+    }
+  } else {
+    formdata.steps["security-review"].status = "CANNOT START YET";
+    completedSections.delete("security");
+  }
+
+  // If every other step is COMPLETED or NOT REQUIRED, set the
+  //  check step to NOT STARTED.
+  const allSteps = new Set(Object.keys(formdata.steps));
+  ["check", "declaration", "confirmation"].forEach((s) => allSteps.delete(s));
+  if (!["check", "declaration", "confirmation"].includes(currentStep)) {
+    completedSections.delete("check");
+    if (everyStepCompleted([...allSteps], formdata)) {
+      formdata.steps["check"].status = "NOT STARTED";
+    } else {
+      formdata.steps["check"].status = "CANNOT START YET";
+    }
+  } else {
+    completedSections.add("check");
+  }
+
+  // Update the number of completed sections
+  formdata.completedSections = completedSections.size;
+};
+
+const everyStepCompleted = (steps: string[], formdata: FormData) => {
+  return steps.every((step) =>
+    ["COMPLETED", "NOT REQUIRED"].includes(formdata.steps[step].status),
+  );
+};
 
 function validateDate(day: number, month: number, year: number): string {
   const errors = new Set<string>();
